Fixed regex issue in port-ipc rest api for /vm-cfg
Port-ipc rest api for /vm-cfg/ is failing, if request
has the dot in the uri. The regex which is used to match
the pod name in /vm-cfg/ does not have the dot. So
contrail-vrouter-agent wont find the match and returns
404. Modified regex to match all substring in /vm-cfg/

Closes-jira-bug: CEM-17010
Change-Id: I9a3cf1d19343767e77142cb5d514b729f6ca7797
(cherry picked from commit cb5161e62ded9e1b236f0ebd3bce39ecba34fb0c)